rename scene TiledMapTest => Rooms
